add Discriminator class
